requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.11.17
click==8.1.7
coverage==7.4.0
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.99.1
fastapi-pagination==0.12.4
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
orjson==3.9.10
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
protobuf==4.21.12
pydantic==1.10.13
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyJWT==2.8.0
pytest==7.4.3
pytest-cov
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
